fix(unit_test): corrected done placement and decreased precision to 5 ms
Because of copying another test to create this one and me forgetting to change the position of the done callback, done was called on every message input event without testing anything.
After I fixed that, I realized that the maximum offset between messages with an average over 4 messages is sometimes up to 5 ms and can probably go even higher on a machine under high load. We have to discuss this internally on how high or low we want this precision to be.